Enable C++ SW flux driver
Enable C++ SW flux computations. These run alongside the F90
computations right now, and we compare fluxes after both calls.
Unfortunately, it looks like the C++ and F90 codes are NOT BFB at the
moment, so answers are not identical, but agree to better than 1e-5
W/m^2. For now, we set a tolerance of 1e-5 on this comparison, but it
would be worthwhile to check if we can make the C++ and F90 codes BFB to
allow for a robust comparison between the new and old RRTMGP drivers.